Moved WithModal and WithLoader to a new folder 'HOCs' and moved LanguageToggleButton to UI/buttons
Переместил HOC-и в новую папку 'HOCs'. А также переместил компонент LanguageToggleButton в UI/buttons
