+++ src/HOCs/WithLoader/WithLoader.module.scss
