+++ src/components/Todo/TodoTask/TodoTaskModal.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import Modal from '../../Modal/Modal';
+import Modal from '../../../HOCs/WithModal/WithModal';
 import ChangeTaskButton from '../../UI/buttons/ChangeTaskButton/ChangeTaskButton';
 import DeleteButton from '../../UI/buttons/DeleteButton/DeleteButton';
 import DoneButton from '../../UI/buttons/DoneButton/DoneButton';
@@ -72,12 +72,12 @@ class TodoTaskModal extends Component {
                     id={this.props.id}
                   />
                   <DoneButton 
-                    isDoneToggle={this.props.isDoneToggle}
+                    isTaskPropertyToggle={this.props.isTaskPropertyToggle}
                     id={this.props.id}
                     isDone={this.props.isDone}
                   />
                   <ImportantButton
-                    isImportantToggle={this.props.isImportantToggle}
+                    isTaskPropertyToggle={this.props.isTaskPropertyToggle}
                     id={this.props.id}
                     isImportant={this.props.isImportant}
                   />
